Fix definition of sbp_message_send

+++ package/sbp_rtcm3_bridge/src/sbp.h
@@ -20,7 +20,7 @@
 #define NO_TIME          0
 
 int sbp_init(sbp_zmq_rx_ctx_t *rx_ctx, sbp_zmq_tx_ctx_t *tx_ctx);
-void sbp_message_send(u8 msg_type, u8 len, u8 *payload, u16 sender_id);
+void sbp_message_send(u16 msg_type, u8 len, u8 *payload, u16 sender_id);
 int sbp_callback_register(u16 msg_type, sbp_msg_callback_t cb, void *context);
 void sbp_base_obs_invalid(double timediff);
